+++ packages/angular/src/sdk.ts
@@ -12,7 +12,6 @@ import {
   dedupeIntegration,
   functionToStringIntegration,
   inboundFiltersIntegration,
-  spanLoggerIntegration,
 } from '@sentry/core';
 import type { Integration } from '@sentry/types';
 import { logger } from '@sentry/utils';
@@ -31,7 +30,7 @@ export function getDefaultIntegrations(): Integration[] {
   // see:
   //  - https://github.com/getsentry/sentry-javascript/issues/5417#issuecomment-1453407097
   //  - https://github.com/getsentry/sentry-javascript/issues/2744
-  const integrations: Integration[] = [
+  return [
     inboundFiltersIntegration(),
     functionToStringIntegration(),
     breadcrumbsIntegration(),
@@ -40,12 +39,6 @@ export function getDefaultIntegrations(): Integration[] {
     dedupeIntegration(),
     httpContextIntegration(),
   ];
-
-  if (IS_DEBUG_BUILD) {
-    integrations.push(spanLoggerIntegration());
-  }
-
-  return integrations;
 }
 
 /**
