+++ packages/browser/src/index.ts
@@ -10,7 +10,6 @@ export {
   extraErrorDataIntegration,
   rewriteFramesIntegration,
   sessionTimingIntegration,
-  spanLoggerIntegration,
 } from '@sentry/core';
 
 export {
