+++ packages/bun/src/index.ts
@@ -120,7 +120,6 @@ export {
   extraErrorDataIntegration,
   rewriteFramesIntegration,
   sessionTimingIntegration,
-  spanLoggerIntegration,
 } from '@sentry/core';
 
 export type { BunOptions } from './types';
