+++ packages/google-cloud-serverless/src/index.ts
@@ -99,7 +99,6 @@ export {
   extraErrorDataIntegration,
   rewriteFramesIntegration,
   sessionTimingIntegration,
-  spanLoggerIntegration,
 } from '@sentry/core';
 
 export { getDefaultIntegrations, init } from './sdk';
