+++ packages/node/src/index.ts
@@ -93,7 +93,6 @@ export {
   extraErrorDataIntegration,
   rewriteFramesIntegration,
   sessionTimingIntegration,
-  spanLoggerIntegration,
   metricsDefault as metrics,
   startSession,
   captureSession,
