+++ packages/node/src/sdk/init.ts
@@ -9,7 +9,6 @@ import {
   inboundFiltersIntegration,
   linkedErrorsIntegration,
   requestDataIntegration,
-  spanLoggerIntegration,
   startSession,
 } from '@sentry/core';
 import { openTelemetrySetupCheck, setOpenTelemetryContextAsyncContextStrategy } from '@sentry/opentelemetry';
@@ -52,7 +51,6 @@ export function getDefaultIntegrations(options: Options): Integration[] {
     functionToStringIntegration(),
     linkedErrorsIntegration(),
     requestDataIntegration(),
-    spanLoggerIntegration(),
     // Native Wrappers
     consoleIntegration(),
     httpIntegration(),
